+++ packages/contracts-bridge/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- (feat): accountant in typechain lib
 - (minor): differentiate EthereumBridgeRouter
 - (minor): add NFT accountant
 - remove ability to handle fast liquidity messages